유형 검사 페이지 일부 병합
+ 파일 구조 변경 및 기타 UI 수정
