+++ client/public/index.html
@@ -9,7 +9,7 @@
 		<title>Escape Velocity by HTML5 UP</title>
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-		<link rel="stylesheet" href="assets/css/main.css" />
+		<!--<link rel="stylesheet" href="assets/css/main.css" />-->
 	</head>
 	<body class="homepage is-preload">
 		<div id="root">
