+++ client/src/assets/css/main.css
@@ -2060,7 +2060,7 @@ input, select, textarea {
 		justify-content: center;
 		position: relative;
 		height: 24em;
-		background: url("../../images/banner.jpg") center center;
+		background: url("images/banner.jpg") center center;
 		background-size: cover;
 		padding: 0;
 	}
