--- client/src/components/Mbti.js
@@ -1,15 +1,0 @@
-import React from "react";
-import {
-  BrowserRouter as Router,
-  Route,
-  Switch,
-  Redirect,
-  Link
-} from "react-router-dom";
-const Mbti = () => (
-    <div>
-      <h1>MBTI 임시 컴포넌트</h1>
-    </div>
-);
-
-export default Mbti;
